Move node classes to separate file
