+++ crates/modelardb_embedded/bindings/python/modelardb/node.py
@@ -1,0 +1,39 @@
+# Copyright 2025 The ModelarDB Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+
+
+@dataclass
+class Server:
+    """A ModelarDB edge or cloud server node.
+
+    :param url: The URL of the ModelarDB server node.
+    :type url: str
+    """
+
+    def __init__(self, url: str):
+        self.url: str = url
+
+
+@dataclass
+class Manager:
+    """A ModelarDB manager node.
+
+    :param url: The URL of the ModelarDB manager node.
+    :type url: str
+    """
+
+    def __init__(self, url: str):
+        self.url: str = url
